langgraph_agent/agent/requirements.txt
langchain
langchain-google-genai
langchain-community
copilotkit
uvicorn
python-dotenv
langchain-core
langgraph-cli
load_dotenv
langgraph-cli[inmem]
